Remove Thrift history client.
