+++ client/clientBean.go
@@ -54,7 +54,6 @@ type (
 	// Bean in an collection of clients
 	Bean interface {
 		GetHistoryClient() history.Client
-		GetHistoryClientGRPC() history.ClientGRPC
 		SetHistoryClient(client history.Client)
 		GetMatchingClient(domainIDToName DomainIDToNameFunc) (matching.Client, error)
 		SetMatchingClient(client matching.Client)
@@ -75,7 +74,6 @@ type (
 		sync.Mutex
 		currentCluster        string
 		historyClient         history.Client
-		historyClientGRPC     history.ClientGRPC
 		matchingClient        atomic.Value
 		remoteAdminClients    map[string]admin.Client
 		remoteFrontendClients map[string]frontend.Client
@@ -108,11 +106,6 @@ type (
 func NewClientBean(factory Factory, dispatcherProvider DispatcherProvider, clusterMetadata cluster.Metadata) (Bean, error) {
 
 	historyClient, err := factory.NewHistoryClient()
-	if err != nil {
-		return nil, err
-	}
-
-	historyClientGRPC, err := factory.NewHistoryClientGRPC()
 	if err != nil {
 		return nil, err
 	}
@@ -150,7 +143,6 @@ func NewClientBean(factory Factory, dispatcherProvider DispatcherProvider, clust
 		currentCluster:        clusterMetadata.GetCurrentClusterName(),
 		factory:               factory,
 		historyClient:         historyClient,
-		historyClientGRPC:     historyClientGRPC,
 		remoteAdminClients:    remoteAdminClients,
 		remoteFrontendClients: remoteFrontendClients,
 	}, nil
@@ -158,10 +150,6 @@ func NewClientBean(factory Factory, dispatcherProvider DispatcherProvider, clust
 
 func (h *clientBeanImpl) GetHistoryClient() history.Client {
 	return h.historyClient
-}
-
-func (h *clientBeanImpl) GetHistoryClientGRPC() history.ClientGRPC {
-	return h.historyClientGRPC
 }
 
 func (h *clientBeanImpl) SetHistoryClient(
