+++ common/resource/resourceImpl.go
@@ -103,7 +103,6 @@ type (
 		matchingClient    matching.Client
 		historyRawClient  history.Client
 		historyClient     history.Client
-		historyClientGRPC history.ClientGRPC
 		clientBean        client.Bean
 
 		// persistence clients
@@ -248,13 +247,6 @@ func New(
 	historyClient := history.NewRetryableClient(
 		historyRawClient,
 		common.CreateHistoryServiceRetryPolicy(),
-		common.IsWhitelistServiceTransientError,
-	)
-
-	historyRawClientGRPC := clientBean.GetHistoryClientGRPC()
-	historyClientGRPC := history.NewRetryableClientGRPC(
-		historyRawClientGRPC,
-		common.CreateHistoryServiceRetryPolicy(),
 		common.IsWhitelistServiceTransientErrorGRPC,
 	)
 
@@ -317,7 +309,6 @@ func New(
 		matchingClient:    matchingClient,
 		historyRawClient:  historyRawClient,
 		historyClient:     historyClient,
-		historyClientGRPC: historyClientGRPC,
 		clientBean:        clientBean,
 
 		// persistence clients
@@ -533,11 +524,6 @@ func (h *Impl) GetHistoryRawClient() history.Client {
 // GetHistoryClient return history client with retry policy
 func (h *Impl) GetHistoryClient() history.Client {
 	return h.historyClient
-}
-
-// GetHistoryClientGRPC return history client with retry policy
-func (h *Impl) GetHistoryClientGRPC() history.ClientGRPC {
-	return h.historyClientGRPC
 }
 
 // GetRemoteAdminClient return remote admin client for given cluster name
