+++ common/resource/resourceTest.go
@@ -31,7 +31,6 @@ import (
 	"go.uber.org/yarpc"
 	"go.uber.org/zap"
 
-	"github.com/temporalio/temporal/.gen/go/history/historyservicetest"
 	"github.com/temporalio/temporal/.gen/proto/adminservicemock"
 	"github.com/temporalio/temporal/.gen/proto/historyservicemock"
 	"github.com/temporalio/temporal/.gen/proto/matchingservicemock"
@@ -84,8 +83,7 @@ type (
 		SDKClient            *workflowservicemock.MockWorkflowServiceClient
 		FrontendClient       *workflowservicemock.MockWorkflowServiceClient
 		MatchingClient       *matchingservicemock.MockMatchingServiceClient
-		HistoryClient        *historyservicetest.MockClient
-		HistoryClientGRPC    *historyservicemock.MockHistoryServiceClient
+		HistoryClient        *historyservicemock.MockHistoryServiceClient
 		RemoteAdminClient    *adminservicemock.MockAdminServiceClient
 		RemoteFrontendClient *workflowservicemock.MockWorkflowServiceClient
 		ClientBean           *client.MockBean
@@ -129,15 +127,13 @@ func NewTest(
 
 	frontendClient := workflowservicemock.NewMockWorkflowServiceClient(controller)
 	matchingClient := matchingservicemock.NewMockMatchingServiceClient(controller)
-	historyClient := historyservicetest.NewMockClient(controller)
-	historyClientGRPC := historyservicemock.NewMockHistoryServiceClient(controller)
+	historyClient := historyservicemock.NewMockHistoryServiceClient(controller)
 	remoteFrontendClient := workflowservicemock.NewMockWorkflowServiceClient(controller)
 	remoteAdminClient := adminservicemock.NewMockAdminServiceClient(controller)
 	clientBean := client.NewMockBean(controller)
 	clientBean.EXPECT().GetFrontendClient().Return(frontendClient).AnyTimes()
 	clientBean.EXPECT().GetMatchingClient(gomock.Any()).Return(matchingClient, nil).AnyTimes()
 	clientBean.EXPECT().GetHistoryClient().Return(historyClient).AnyTimes()
-	clientBean.EXPECT().GetHistoryClientGRPC().Return(historyClientGRPC).AnyTimes()
 	clientBean.EXPECT().GetRemoteAdminClient(gomock.Any()).Return(remoteAdminClient).AnyTimes()
 	clientBean.EXPECT().GetRemoteFrontendClient(gomock.Any()).Return(remoteFrontendClient).AnyTimes()
 
@@ -198,7 +194,6 @@ func NewTest(
 		FrontendClient:       frontendClient,
 		MatchingClient:       matchingClient,
 		HistoryClient:        historyClient,
-		HistoryClientGRPC:    historyClientGRPC,
 		RemoteAdminClient:    remoteAdminClient,
 		RemoteFrontendClient: remoteFrontendClient,
 		ClientBean:           clientBean,
@@ -351,11 +346,6 @@ func (s *Test) GetHistoryRawClient() history.Client {
 // GetHistoryClient for testing
 func (s *Test) GetHistoryClient() history.Client {
 	return s.HistoryClient
-}
-
-// GetHistoryClientGRPC for testing
-func (s *Test) GetHistoryClientGRPC() history.ClientGRPC {
-	return s.HistoryClientGRPC
 }
 
 // GetRemoteAdminClient for testing
