+++ service/frontend/adminHandler_test.go
@@ -86,7 +86,7 @@ func (s *adminHandlerSuite) SetupTest() {
 	s.controller = gomock.NewController(s.T())
 	s.mockResource = resource.NewTest(s.controller, metrics.Frontend)
 	s.mockDomainCache = s.mockResource.DomainCache
-	s.mockHistoryClient = s.mockResource.HistoryClientGRPC
+	s.mockHistoryClient = s.mockResource.HistoryClient
 	s.mockHistoryV2Mgr = s.mockResource.HistoryMgr
 
 	params := &service.BootstrapParams{
